reformat, moved query sets to models.py
