+++ src/homework/models/__init__.py
@@ -1,9 +1,7 @@
-
-from homework.models.question import Question
-
 from homework.models.answer import Answer
 from homework.models.answer_access_log_entry import AnswerAccessLogEntry
 from homework.models.answer_cross_check import AnswerCrossCheck
+from homework.models.question import Question
 
 __all__ = [
     'Answer',
